+++ src/NinaHA.Plugin/SequenceItems/SequenceItemTemplates.xaml.cs
@@ -1,0 +1,13 @@
+using System.ComponentModel.Composition;
+using System.Windows;
+
+namespace NinaHA.Plugin.SequenceItems {
+
+    [Export(typeof(ResourceDictionary))]
+    public partial class SequenceItemTemplates : ResourceDictionary {
+
+        public SequenceItemTemplates() {
+            InitializeComponent();
+        }
+    }
+}